Split file parsing and game loading
